parkauto/requirements.txt
asgiref==3.9.1
attrs==25.3.0
bcrypt==4.3.0
channels==4.2.2
channels_redis==4.3.0
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.9.0
